Adjusting the mod.rs files in the modules
