+++ src/laser/mod.rs
@@ -36,40 +36,15 @@ impl<'a> System<'a> for AttachLaserComponentsToNewlyCreatedAtomsSystem {
 			);
 			updater.insert(
 				ent,
-				doppler::DopplerShiftSamplers {
-					contents: [doppler::DopplerShiftSampler::default(); BEAM_LIMIT],
-				},
-			);
-			updater.insert(
-				ent,
 				intensity::LaserIntensitySamplers {
 					contents: [intensity::LaserIntensitySampler::default(); BEAM_LIMIT],
 				},
 			);
 			updater.insert(
 				ent,
-				sampler::LaserDetuningSamplers {
-					contents: [sampler::LaserDetuningSampler::default(); BEAM_LIMIT],
-				},
-			);
-			updater.insert(
-				ent,
-				rate::RateCoefficients {
-					contents: [rate::RateCoefficient::default(); BEAM_LIMIT],
-				},
-			);
-			updater.insert(ent, twolevel::TwoLevelPopulation::default());
-			updater.insert(ent, photons_scattered::TotalPhotonsScattered::default());
-			updater.insert(
-				ent,
-				photons_scattered::ExpectedPhotonsScatteredVector {
-					contents: [photons_scattered::ExpectedPhotonsScattered::default(); BEAM_LIMIT],
-				},
-			);
-			updater.insert(
-				ent,
-				photons_scattered::ActualPhotonsScatteredVector {
-					contents: [photons_scattered::ActualPhotonsScattered::default(); BEAM_LIMIT],
+				intensity_gradient::LaserIntensityGradientSamplers {
+					contents: [intensity_gradient::LaserIntensityGradientSampler::default();
+						BEAM_LIMIT],
 				},
 			);
 		}
@@ -90,16 +65,6 @@ pub fn add_systems_to_dispatch(builder: &mut DispatcherBuilder<'static, 'static>
 		deps,
 	);
 	builder.add(
-		cooling::AttachIndexToCoolingLightSystem,
-		"attach_cooling_index",
-		deps,
-	);
-	builder.add(
-		cooling::IndexCoolingLightsSystem,
-		"index_cooling_lights",
-		deps,
-	);
-	builder.add(
 		sampler::InitialiseLaserSamplerMasksSystem,
 		"initialise_laser_sampler_masks",
 		deps,
@@ -115,68 +80,14 @@ pub fn add_systems_to_dispatch(builder: &mut DispatcherBuilder<'static, 'static>
 		&["index_cooling_lights", INTEGRATE_POSITION_SYSTEM_NAME],
 	);
 	builder.add(
-		doppler::CalculateDopplerShiftSystem,
-		"calculate_doppler_shift",
-		&["index_cooling_lights"],
-	);
-	builder.add(
-		sampler::CalculateLaserDetuningSystem,
-		"calculate_laser_detuning",
-		&[
-			"calculate_doppler_shift",
-			"zeeman_shift",
-			"index_cooling_lights",
-		],
-	);
-	builder.add(
-		rate::CalculateRateCoefficientsSystem,
-		"calculate_rate_coefficients",
-		&["calculate_laser_detuning"],
-	);
-	builder.add(
-		twolevel::CalculateTwoLevelPopulationSystem,
-		"calculate_twolevel",
-		&["calculate_rate_coefficients", "fill_laser_sampler_masks"],
-	);
-	builder.add(
-		photons_scattered::CalculateMeanTotalPhotonsScatteredSystem,
-		"calculate_total_photons",
-		&["calculate_twolevel"],
-	);
-	builder.add(
-		photons_scattered::CalculateExpectedPhotonsScatteredSystem,
-		"calculate_expected_photons",
-		&["calculate_total_photons", "fill_laser_sampler_masks"],
-	);
-	builder.add(
-		photons_scattered::CalculateActualPhotonsScatteredSystem,
-		"calculate_actual_photons",
-		&["calculate_expected_photons"],
-	);
-	builder.add(
-		force::CalculateAbsorptionForcesSystem,
-		"calculate_absorption_forces",
-		&["calculate_actual_photons", INTEGRATE_POSITION_SYSTEM_NAME],
-	);
-	builder.add(
-		repump::RepumpSystem,
-		"repump",
-		&["calculate_absorption_forces"],
-	);
-	builder.add(
-		force::ApplyEmissionForceSystem,
-		"calculate_emission_forces",
-		&[
-			"calculate_absorption_forces",
-			INTEGRATE_POSITION_SYSTEM_NAME,
-		],
+		intensity_gradient::SampleLaserIntensityGradientSystem,
+		"sample_intensity_gradient",
+		&["index_dipole_lights"],
 	);
 }
 
 /// Registers resources required by magnetics to the ecs world.
 pub fn register_components(world: &mut World) {
-	world.register::<cooling::CoolingLight>();
-	world.register::<cooling::CoolingLightIndex>();
 	world.register::<gaussian::GaussianBeam>();
 	world.register::<gaussian::CircularMask>();
 	world.register::<gaussian::GaussianReferenceFrame>();
