+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod initiate;
 pub mod integration_tests;
 pub mod integrator;
 pub mod laser;
+pub mod laser_cooling;
 pub mod magnetic;
 pub mod maths;
 pub mod output;
